Extract safety consumer to its own module
Extract the Safety dialect's AST consumer to its own module to make the
project's code more modular.
