+++ include/macroni/ASTConsumers/Safety/SafetyASTConsumer.hpp
